Patch the failing PostgreSQL integration tests
Curiously, the Projection extension spec started failing in GitHub
Actions' pipelines out of the blue, with "unable to connect to
localhost" errors. The log files, unfortunately, do not provide enough
information to suggest what might be going on (neither the Reactor nor
the Postgres ones), so I am going to leave the investigation for
another day.